+++ app/src/main/java/com/example/misradbru/realdeal/data/ProductRepositoryImpl.java
@@ -1,13 +1,24 @@
 package com.example.misradbru.realdeal.data;
 
+import android.content.Context;
 import android.support.annotation.NonNull;
 import android.util.Log;
+import android.widget.ListView;
 
+import com.example.misradbru.realdeal.products.ProductsAdapter;
+import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.FirebaseFirestore;
 import com.google.firebase.firestore.FirebaseFirestoreSettings;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
+import com.google.firebase.firestore.QuerySnapshot;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class ProductRepositoryImpl implements ProductRepository {
     private final String TAG = "ProductRepositoryImpl";
@@ -36,6 +47,33 @@ public class ProductRepositoryImpl implements ProductRepository {
                     @Override
                     public void onFailure(@NonNull Exception e) {
                         Log.w(TAG, "Error adding document", e);
+                    }
+                });
+    }
+
+    @Override
+    public void getProducts(FirebaseAuth mAuth, Context context, final ListView mProductListView) {
+        final List<Product> mProductList = new ArrayList<>();
+        final ProductsAdapter mProductsAdapter = new ProductsAdapter(context, mProductList);
+        FirebaseFirestore db = FirebaseFirestore.getInstance();
+        String userId = mAuth.getUid();
+        db.collection("products")
+                .whereEqualTo("uid", userId)
+                .get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if(task.isSuccessful()){
+                            for(QueryDocumentSnapshot document : task.getResult()){
+                                Log.d(TAG, document.getId() + " => " + document.getData());
+                                Product product = document.toObject(Product.class);
+                                mProductList.add(product);
+                            }
+                            mProductListView.setAdapter(mProductsAdapter);
+                        }
+                        else {
+                            Log.d(TAG, "Error getting documents: ", task.getException());
+                        }
                     }
                 });
     }
