+++ app/src/main/res/layout/activity_main.xml
@@ -6,7 +6,6 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     tools:context=".products.MainActivity">
-
     <LinearLayout
         android:layout_width="match_parent"
         android:orientation="vertical"
